found a mistake
some huge changes in refactoring
