+++ helpers/ctrlWrapper.js
@@ -1,8 +1,11 @@
+
+
 const ctrlWrapper = ctrl => {
-    const func = async (req,res,next) => {
+    const func = async(req, res, next) => {
         try {
-            await ctrl(req,res,next);
-        } catch (error) {
+            await ctrl(req, res, next);
+        }
+        catch(error) {
             next(error);
         }
     }
