+++ middleware/index.js
@@ -1,4 +1,5 @@
 const validateBody = require("./validateBody");
 
 module.exports = {
-    validateBody, };
+    validateBody,
+}
