+++ models/contacts.json
@@ -48,20 +48,14 @@
     "phone": "(233) 738-2360"
   },
   {
-    "id": "e6ywwRe4jcqxXfCZOj_1e",
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993"
+    "contactId": "Y0Or_Qem9g0_XezGeRRO9",
+    "name": "max",
+    "email": "oleksii@gmail.com",
+    "phone": "2222222221"
   },
   {
-    "id": "rsKkOQUi80UsgVPCcLZZW",
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
-  },
-  {
-    "contactId": "RDY1WOmP15li2IIsHOjoc",
-    "name": "Oleksii Panaioti",
+    "contactId": "kixH0gzWwyPOkHUorA14L",
+    "name": "max",
     "email": "oleksii@gmail.com",
     "phone": "2222222221"
   }
