+++ routes/api/contacts.js
@@ -2,13 +2,13 @@ const express = require("express");
 
 const ctrl = require("../../controllers/contacts");
 
+const {validateBody} = require("../../middleware");
+
+const schemas = require("../../shemas/contacts");
+
 const router = express.Router();
 
-const {validateBody} = require("../../middlewares/validateBody");
-
-const schemas = require("../../schemas/contacts");
-
-router.get("/", ctrl.getAll );
+router.get("/", ctrl.getAll);
 
 router.get("/:id", ctrl.getById);
 
@@ -16,6 +16,6 @@ router.post("/", validateBody(schemas.addSchema), ctrl.add);
 
 router.delete("/:id", ctrl.deleteById);
 
-router.put("/:id",validateBody(schemas.addSchema), ctrl.updateById);
+router.put("/:id", validateBody(schemas.addSchema),  ctrl.updateById);
 
 module.exports = router;
